Arduino board test added
