+++ code/firmware/test/Header.hpp
@@ -2,6 +2,9 @@
 #define HEADER_HPP
 
 // Declarations
-int addsometothis();
+void test_led_builtin_pin_number(void);
+void test_led_state_high(void);
+void test_led_state_low(void);
+
 
 #endif
