+++ code/firmware/test/README
@@ -1,4 +1,4 @@
-# Instructions for adding tests :
+ Instructions for adding tests :
 
 1. Create separate test files for each item with functions
 2. Include above function declarations in Header.hpp
