+++ code/firmware/test/test_ArduinoUNO.cpp
@@ -28,33 +28,33 @@ void test_led_state_low(void)
   TEST_ASSERT_EQUAL(LOW, digitalRead(LED_BUILTIN));
 }
 
-void setup()
-{
-  // NOTE!!! Wait for >2 secs
-  // if board doesn't support software reset via Serial.DTR/RTS
-  delay(2000);
+// void setup()
+// {
+//   // NOTE!!! Wait for >2 secs
+//   // if board doesn't support software reset via Serial.DTR/RTS
+//   delay(2000);
 
-  pinMode(LED_BUILTIN, OUTPUT);
+//   pinMode(LED_BUILTIN, OUTPUT);
 
-  UNITY_BEGIN(); // IMPORTANT LINE!
-  RUN_TEST(test_led_builtin_pin_number);
-}
+//   UNITY_BEGIN(); // IMPORTANT LINE!
+//   RUN_TEST(test_led_builtin_pin_number);
+// }
 
-uint8_t i = 0;
-uint8_t max_blinks = 5;
+// uint8_t i = 0;
+// uint8_t max_blinks = 5;
 
-void loop()
-{
-  if (i < max_blinks)
-  {
-    RUN_TEST(test_led_state_high);
-    delay(500);
-    RUN_TEST(test_led_state_low);
-    delay(500);
-    i++;
-  }
-  else if (i == max_blinks)
-  {
-    UNITY_END(); // stop unit testing
-  }
-}
+// void loop()
+// {
+//   if (i < max_blinks)
+//   {
+//     RUN_TEST(test_led_state_high);
+//     delay(500);
+//     RUN_TEST(test_led_state_low);
+//     delay(500);
+//     i++;
+//   }
+//   else if (i == max_blinks)
+//   {
+//     UNITY_END(); // stop unit testing
+//   }
+// }
